+++ app/_components/_nav/mypage.tsx
@@ -1,3 +1,4 @@
+import cx from 'clsx';
 import Image from 'next/image';
 import Link from 'next/link';
 
@@ -25,7 +26,7 @@ export default function MypageNav({ activePage, className }: IMypageNavProps) {
   if (!navItem) return null;
 
   return (
-    <div className={className || 'fixed z-20 top-0 left-[50%] -translate-x-1/2 w-[800px] h-[215px]'}>
+    <div className={cx('w-[800px] h-[215px]', className)}>
       <Image src={navItem.background} alt={navItem.title} width={800} height={215} />
       <div className="absolute top-[60px] left-0 w-[800px] h-[150px] flex flex-row items-start z-10">
         <Link href={URLS.getMypagePageByKind('studio')} className="w-6/12 h-full" />
