+++ app/_libs/api/artwork.ts
@@ -1,6 +1,6 @@
 import { createHttpClient } from '../http-client';
 
-import type { IArtwork, IArtworkDetail } from '@/app/_types/artwork';
+import type { IArtworkDetail, IArtworkReward } from '@/app/_types/artwork';
 import { URLS } from '@/app/constants';
 
 export async function fetchArtworkPuzzles(uid: string): Promise<IArtworkDetail> {
@@ -8,8 +8,8 @@ export async function fetchArtworkPuzzles(uid: string): Promise<IArtworkDetail> 
   return client.get<IArtworkDetail>(URLS.fetchArtworkPuzzlesByUid(uid));
 }
 
-export async function fetchArtworkCompleted(): Promise<IArtwork[]> {
+export async function fetchArtworkCompleted(): Promise<IArtworkReward[]> {
   const client = await createHttpClient();
 
-  return client.get<IArtwork[]>(URLS.fetchArtworkCompleted());
+  return client.get<IArtworkReward[]>(URLS.fetchArtworkCompleted());
 }
