+++ app/_types/artwork.ts
@@ -5,10 +5,19 @@ export interface IArtwork {
   artworkSeq: number;
   artworkTitle: string;
   artworkUid: string;
+  imageUrl: string;
 }
 
 export interface IArtworkDetail {
   artwork: IArtwork;
   email: string; // 로그인 여부
   puzzles: IPuzzle[];
+}
+
+export interface IArtworkReward {
+  artworkImgUrl: string;
+  artworkSeq: number;
+  artworkUid: string;
+  rewardCode: string;
+  rewardImgUrl: string;
 }
